create trip from create screen
