+++ apps/app/src/domain/Trip/Models/Trips.ts
@@ -1,0 +1,5 @@
+export enum Trips {
+  Energy = "Energy",
+  Psy = "Psy",
+  Love = "Love",
+}
